将更新podspec内容抽离出来
